wrap city dao by repository

+++ app/src/main/java/com/danikula/sunny/web/dto/PlacesSearchResponse.kt
@@ -4,7 +4,7 @@ import com.danikula.sunny.model.City
 
 data class PlacesSearchResponse(val list: List<Place>)
 
-data class Place(val id: Int, val name: String, val sys: PlaceCountry) {
+data class Place(val id: Long, val name: String, val sys: PlaceCountry) {
     fun toCity(): City {
         return City(id, name, sys.country)
     }
